requirement.txt
absl-py==0.12.0
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.2.0
audioread==3.0.0
beautifulsoup4==4.12.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.0
chardet==5.0.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.4
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
Django==3.2.13
django-cors-headers==3.13.0
django-environ==0.8.1
django-sslserver==0.22
djangorestframework==3.13.1
dlib==19.24.1
face-recognition==1.3.0
face-recognition-models==0.3.0
ffmpeg-python==0.2.0
filelock==3.0.12
flatbuffers==2.0
fluidsynth==0.2
fonttools==4.20.0
frozenlist==1.3.0
future==0.18.3
gast==0.3.3
google==3.0.0
google-api-core==2.11.0
google-api-python-client==2.20.0
google-auth==2.7.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-cloud-core==2.3.2
google-cloud-speech==2.7.0
google-cloud-storage==1.45.0
google-cloud-vision==2.3.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.0.0
googleapis-common-protos==1.53.0
grpcio==1.39.0
grpcio-status==1.42.0
gTTS==2.2.2
gunicorn==20.1.0
h5py==3.3.0
httplib2==0.20.1
huggingface-hub==0.0.17
idna==3.4
importlib-metadata==4.8.1
importlib-resources==5.3.0
jax==0.2.19
joblib==1.0.1
jsonpickle==2.0.0
keras==2.6.0
kiwisolver==1.3.1
libclang==11.0.1
libcst==0.3.21
librosa==0.8.1
llvmlite==0.37.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.5.1
midi2audio==0.1.3
MIDIUtil==1.2.1
ml-dtypes==0.2.0
more-itertools==8.10.0
mpi4py==3.1.0
multidict==6.0.2
music21==6.7.1
mypy-extensions==0.4.3
numba==0.53.1
numpy==1.20.3
oauthlib==3.2.0
openai==0.18.0
opencv-python==4.5.4.58
opencv-python-headless==4.5.4.58
opt-einsum==3.3.0
packaging==21.3
Pillow==9.0.0
proto-plus==1.19.5
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydub==0.25.1
PyJWT==2.2.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.4
PyYAML==5.4.1
regex==2023.11.1
requests==2.26.0
requests-oauthlib==1.3.1
resampy==0.2.2
rsa==4.7.2
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
soundfile==0.11.0
soupsieve==2.3.1
sqlparse==0.4.2
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorflow==2.6.0
tensorflow-estimator==2.6.0
tensorflow-intel==2.6.0
tensorflow-io-gcs-filesystem==0.21.0
termcolor==1.1.0
threadpoolctl==3.0.0
tokenizers==0.10.3
torch==1.8.2+cpu
tqdm==4.62.3
transformers==4.12.2
typing-inspect==0.7.1
typing_extensions==4.1.1
uritemplate==3.0.1
urllib3==1.26.7
webcolors==1.11.1
Werkzeug==2.0.2
wrapt==1.12.1
yarl==1.7.0
zipp==3.7.0